refactor(JOE-936): split environments into util, lifecycle, backends
Extract acyclic leaves under environments/ (util/redaction, lifecycle,
local-container, remote-crabbox). Façade environments.ts ~2275→~1262 LOC;
hard budget 2320→1300.
